# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BoardSight API
  version: 0.1.0
  description: >
    BoardSight Chess backend — game analysis, move commentary, and
    real-time multiplayer relay.

servers:
  - url: http://localhost:8000
    description: Local development server

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    # --- Analysis -----------------------------------------------------------
    AnalysisRequest:
      type: object
      required:
        - pgn
      properties:
        pgn:
          type: string
          description: PGN string of the game to analyse.
          example: "1. e4 e5 2. Nf3 Nc6 3. Bb5 a6 1-0"
        depth:
          type: integer
          default: 18
          description: Stockfish search depth.
        include_llm_takeaways:
          type: boolean
          default: false
          description: Whether to append AI-generated narrative takeaways.
        api_key:
          type: string
          nullable: true
          description: Optional OpenAI / Anthropic key forwarded to the LLM worker.

    AnalysisJobCreated:
      type: object
      required:
        - job_id
        - status
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]

    AnalysisJob:
      type: object
      required:
        - job_id
        - status
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, done, error]
        result:
          type: object
          nullable: true
          description: Full analysis payload; present only when status is "done".
        error:
          type: string
          nullable: true
          description: Error message; present only when status is "error".

    AnalysisResult:
      type: object
      required:
        - result
      properties:
        result:
          type: object
          description: Full analysis payload returned by the Stockfish worker.

    AnalysisPending:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [pending]

    # --- Commentary ---------------------------------------------------------
    CommentaryRequest:
      type: object
      required:
        - fen
        - move
      properties:
        fen:
          type: string
          description: FEN string representing the board position.
        move:
          type: string
          description: The move to comment on (SAN or UCI).

    CommentaryResponse:
      type: object
      required:
        - comment
      properties:
        comment:
          type: string
          description: Natural-language comment generated by the LLM.

    # --- Health -------------------------------------------------------------
    HealthResponse:
      type: object
      required:
        - status
        - version
      properties:
        status:
          type: string
          example: ok
        version:
          type: string
          example: "1.0.0"

    HealthReadyResponse:
      type: object
      required:
        - db
        - redis
        - status
      properties:
        db:
          type: boolean
          description: Whether Postgres is reachable.
        redis:
          type: boolean
          description: Whether Redis is reachable.
        status:
          type: string
          enum: [ready, degraded]
          description: "`ready` when both dependencies are up; `degraded` otherwise."

    # --- Errors -------------------------------------------------------------
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:

  /health:
    get:
      operationId: health_check
      summary: Liveness check
      description: Always returns 200 while the process is running.
      tags: [health]
      responses:
        "200":
          description: Service is up.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                version: "1.0.0"

  /health/ready:
    get:
      operationId: health_ready
      summary: Readiness check
      description: >
        Verifies that both backing dependencies (Postgres and Redis) are
        reachable.

        - `200` — both `db` and `redis` are up; `status` is `"ready"`.
        - `503` — at least one dependency is unreachable; `status` is
          `"degraded"`. Inspect the `db` and `redis` boolean fields to
          identify which service is down.
      tags: [health]
      responses:
        "200":
          description: All dependencies are healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthReadyResponse"
              example:
                db: true
                redis: true
                status: ready
        "503":
          description: One or more dependencies are unreachable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthReadyResponse"
              example:
                db: false
                redis: false
                status: degraded

  /v1/analysis/jobs:
    post:
      operationId: submit_analysis_job
      summary: Submit a game for analysis
      tags: [analysis]
      security:
        - ApiKeyHeader: []
        - {}   # also allow unauthenticated (key passed in body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisRequest"
      responses:
        "201":
          description: Job accepted and queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJobCreated"
        "422":
          description: Validation error (missing / malformed fields).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /v1/analysis/jobs/{job_id}:
    get:
      operationId: get_analysis_job
      summary: Poll job status and result
      description: >
        Returns the job's current status. When `status` is `"done"`, the
        `result` field contains the full analysis payload. When `status` is
        `"error"`, the `error` field contains the failure reason.
      tags: [analysis]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Job found — status may be queued, running, done, or error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJob"
        "404":
          description: Job not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /v1/analysis/jobs/{job_id}/result:
    get:
      operationId: get_analysis_result
      summary: Fetch the full analysis result payload
      description: >
        Returns **only** the result object once the job is complete.

        - `200` — job is done; body is `{"result": {...}}`.
        - `202` — job is still queued or running; body is `{"status": "pending"}`.
        - `404` — unknown job_id.
      tags: [analysis]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Analysis complete — full result payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "202":
          description: Job still in progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisPending"
        "404":
          description: Job not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /v1/commentary:
    post:
      operationId: get_commentary
      summary: Generate a natural-language comment for a move
      tags: [commentary]
      security:
        - ApiKeyHeader: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentaryRequest"
      responses:
        "200":
          description: Comment generated (or canned fallback).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentaryResponse"
        "422":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /ws/relay/{code}:
    get:
      operationId: websocket_relay
      summary: Real-time multiplayer relay (WebSocket upgrade)
      description: >
        Upgrade to a WebSocket connection to participate in a relay session.

        **Roles** (query param `role`):

        | Value | Description |
        |-------|-------------|
        | `host` | Creates / owns the session; sends and receives moves. Only one host per code. |
        | `guest` | Joins an existing session; sends and receives moves. Only one guest per code. |
        | `spectate` | Read-only observer; receives all forwarded frames but cannot send. |

        **Client → Server frames** (JSON):

        | type | extra fields |
        |------|-------------|
        | `MOVE` | `move: {from, to, promotion?}`, `clockState` |
        | `CLOCK_SYNC` | `clockState` |
        | `CORRECTION_REQUEST` | — |
        | `CORRECTION_APPROVED` | `fen: string` |
        | `CORRECTION_DENIED` | — |
        | `GAME_OVER` | `result: string` |

        **Server → Client frames** (JSON):

        | type | description |
        |------|-------------|
        | `peer_joined` | Other player connected (or both connected, for spectators) |
        | `peer_disconnected` | Other player disconnected |
        | `<any client frame>` | Forwarded verbatim from the peer |

        **Close codes**:

        | code | reason |
        |------|--------|
        | `4000` | Invalid or missing `role` query param |
        | `4001` | Requested slot (host / guest) already occupied |
      tags: [relay]
      parameters:
        - name: code
          in: path
          required: true
          description: Six-character uppercase session code.
          schema:
            type: string
            pattern: "^[A-Z0-9]{6}$"
            example: "ABC123"
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [host, guest, spectate]
      responses:
        "101":
          description: WebSocket handshake accepted — connection upgraded.
        "400":
          description: Missing or invalid `role` parameter (connection refused before upgrade).